# app/Docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sports Manager API",
        "description": "API para gerenciar dados esportivos, incluindo jogadores, times e jogos.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login de um usuário",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@admin.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "admin.1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/players": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Lista todos os jogadores",
                "operationId": "a355f9879160729f824bdea53e739b05",
                "responses": {
                    "200": {
                        "description": "Lista de jogadores retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "first_name": {
                                                "type": "string",
                                                "example": "Jaylen"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Adams"
                                            },
                                            "position": {
                                                "type": "string",
                                                "example": "G"
                                            },
                                            "height": {
                                                "type": "string",
                                                "example": "6-0"
                                            },
                                            "weight": {
                                                "type": "string",
                                                "example": "225"
                                            },
                                            "college": {
                                                "type": "string",
                                                "example": "St. Bonaventure"
                                            },
                                            "country": {
                                                "type": "string",
                                                "example": "USA"
                                            },
                                            "team": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "conference": {
                                                        "type": "string",
                                                        "example": "East"
                                                    },
                                                    "division": {
                                                        "type": "string",
                                                        "example": "Southeast"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Atlanta"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Hawks"
                                                    },
                                                    "full_name": {
                                                        "type": "string",
                                                        "example": "Atlanta Hawks"
                                                    },
                                                    "abbreviation": {
                                                        "type": "string",
                                                        "example": "ATL"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Cria um novo jogador",
                "operationId": "4e2752a0305de0adfa2a8c9ac1aa37a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "position",
                                    "team_id"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Fred"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "F"
                                    },
                                    "height": {
                                        "type": "string",
                                        "example": "6.5"
                                    },
                                    "weight": {
                                        "type": "string",
                                        "example": "220"
                                    },
                                    "college": {
                                        "type": "string",
                                        "example": "Duke"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "USA"
                                    },
                                    "team_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jogador criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Doe"
                                        },
                                        "position": {
                                            "type": "string",
                                            "example": "Forward"
                                        },
                                        "height": {
                                            "type": "string",
                                            "example": "6.5"
                                        },
                                        "weight": {
                                            "type": "string",
                                            "example": "220"
                                        },
                                        "college": {
                                            "type": "string",
                                            "example": "Duke"
                                        },
                                        "country": {
                                            "type": "string",
                                            "example": "USA"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/players/{id}": {
            "put": {
                "tags": [
                    "Players"
                ],
                "summary": "Atualiza um jogador existente",
                "operationId": "b0080e3337c0615ba30865afac7ad398",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do jogador a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "position"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Jane"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Smith"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Guard"
                                    },
                                    "height": {
                                        "type": "string",
                                        "example": "5.9"
                                    },
                                    "weight": {
                                        "type": "string",
                                        "example": "150"
                                    },
                                    "college": {
                                        "type": "string",
                                        "example": "UCLA"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "USA"
                                    },
                                    "team_id": {
                                        "type": "integer",
                                        "example": "3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jogador atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Jane"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Smith"
                                        },
                                        "position": {
                                            "type": "string",
                                            "example": "Guard"
                                        },
                                        "height": {
                                            "type": "string",
                                            "example": "5.9"
                                        },
                                        "weight": {
                                            "type": "string",
                                            "example": "150"
                                        },
                                        "college": {
                                            "type": "string",
                                            "example": "UCLA"
                                        },
                                        "country": {
                                            "type": "string",
                                            "example": "USA"
                                        },
                                        "team_id": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Players"
                ],
                "summary": "Deleta um jogador existente",
                "operationId": "a0d43e3c36f501bca2882e19d853c171",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do jogador a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jogador deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Player deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jogador não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Players",
            "description": "Players"
        }
    ]
}